add GetUnreadUserNews tests
