+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/NewsArticles/Application/UseCases/NewsArticleWriteOnlyRepositoryTestDoubles.cs
@@ -1,4 +1,4 @@
-﻿namespace DfE.GIAP.Core.UnitTests.NewsArticles.UseCases;
+﻿namespace DfE.GIAP.Core.UnitTests.NewsArticles.Application.UseCases;
 internal static class NewsArticleWriteOnlyRepositoryTestDoubles
 {
     internal static Mock<INewsArticleWriteOnlyRepository> Default() => CreateMock();
