+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/NewsArticles/Application/UseCases/UpdateNewsArticle/UpdateNewsArticleUseCaseTests.cs
@@ -2,8 +2,9 @@
 using DfE.GIAP.Core.Common.CrossCutting;
 using DfE.GIAP.Core.NewsArticles.Application.UseCases.UpdateNewsArticle;
 using DfE.GIAP.Core.SharedTests.TestDoubles;
+using DfE.GIAP.Core.UnitTests.NewsArticles.Application.UseCases;
 
-namespace DfE.GIAP.Core.UnitTests.NewsArticles.UseCases.UpdateNewsArticle;
+namespace DfE.GIAP.Core.UnitTests.NewsArticles.Application.UseCases.UpdateNewsArticle;
 public sealed class UpdateNewsArticleUseCaseTests
 {
     [Theory]
