+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/NewsArticles/Infrastructure/Repositories/CosmosNewsArticleReadRepositoryTests.cs
@@ -3,7 +3,7 @@ using DfE.GIAP.Core.Common.CrossCutting;
 using DfE.GIAP.Core.NewsArticles.Application.Enums;
 using DfE.GIAP.Core.NewsArticles.Infrastructure.Repositories;
 using DfE.GIAP.Core.SharedTests.TestDoubles;
-using DfE.GIAP.Core.UnitTests.NewsArticles.UseCases;
+using DfE.GIAP.Core.UnitTests.NewsArticles.Application.UseCases;
 using DfE.GIAP.Core.UnitTests.TestDoubles;
 
 namespace DfE.GIAP.Core.UnitTests.NewsArticles.Infrastructure.Repositories;
